Add environment class
